% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextAsDataFrame.R
\name{contextAsDataFrame}
\alias{contextAsDataFrame}
\title{Coerce Context DNAStringSet as Data Frame}
\usage{
contextAsDataFrame(vcf, context)
}
\arguments{
\item{vcf}{ExpandedVcf object}

\item{context}{\code{DNAStringSet} returned by \code{snvtyping::getSnvContext} function}
}
\value{
Data frame with columns \code{up} \code{ref} \code{alt} \code{down} \code{context}
\code{snv.context} \code{snv}
}
\description{
This function converts the context DNAStringSet of SNVs to a data frame.
The SNV information is redundant, given their reverse complement, so
they are compressed to only T and C references.
}
\examples{
library(snvtyping)
library(S4Vectors)
context.len <- 3
genome <- BSgenome.Hsapiens.UCSC.hg19::Hsapiens
fl <- system.file("extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
                  mustWork=TRUE)
vcf <- VariantAnnotation::readVcf(fl)
vcf <- VariantAnnotation::expand(vcf)
vcf <- preprocessVcf(vcf)
vcf <- vcf[1:50]
context <- getSnvContext(vcf, genome, context.len)
df <- contextAsDataFrame(vcf, context)
}
