% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnvContext.R
\name{getSnvContext}
\alias{getSnvContext}
\title{Get Genomic Context}
\usage{
getSnvContext(vcf, genome, context.len = 3)
}
\arguments{
\item{vcf}{ExpandedVcf object}

\item{genome}{Set of sequences on which query is made}

\item{context.len}{Length of the genomic context to query. Odd integer number
equal or greater than 3. (Default: 3)}
}
\value{
Genomic contexts as DNAStringSet object of all the SNVs in the
ExpandedVCF
}
\description{
This function returns the genomic context of a SNV, of a defined base-pair
length, from parameters: chromosome, position, reference nucleotide and
alternative nucleotide.
The SNV information is redundant, given their reverse complement, so
they are compressed to only T and C references.
}
\examples{
library(snvtyping)
library(S4Vectors)
context.len <- 3
genome <- BSgenome.Hsapiens.UCSC.hg19::Hsapiens
fl <- system.file("extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
                  mustWork=TRUE)
vcf <- VariantAnnotation::readVcf(fl)
vcf <- VariantAnnotation::expand(vcf)
vcf <- preprocessVcf(vcf)
vcf <- vcf[1:50]
context <- getSnvContext(vcf, genome, context.len)
}
