% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnvCountTable.R
\name{getSnvCountTable}
\alias{getSnvCountTable}
\title{Get Count Table of SNVs}
\usage{
getSnvCountTable(vcf, context, type = "context")
}
\arguments{
\item{vcf}{ExpandedVcf object}

\item{context}{\code{DNAStringSet} returned by \code{snvtyping::getSnvContext} function}

\item{type}{Flag for getting a different
count table: \code{"context"} return a count table for every SNV context,
\code{"context.by.snv"} returns a count table for every reference context grouped
by their SNV type}
}
\value{
Count table of SNVs
}
\description{
Get Count Table of SNVs
}
\examples{
library(snvtyping)
library(S4Vectors)
context.len <- 3
genome <- BSgenome.Hsapiens.UCSC.hg19::Hsapiens
fl <- system.file("extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
                  mustWork=TRUE)
vcf <- VariantAnnotation::readVcf(fl)
vcf <- VariantAnnotation::expand(vcf)
vcf <- preprocessVcf(vcf)
vcf <- vcf[1:50]
context <- getSnvContext(vcf, genome, context.len)
getSnvCountTable(vcf, context, type = "context")
getSnvCountTable(vcf, context, type = "context.by.snv")
}
