% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSnvTypes.R
\name{plotSnvTypes}
\alias{plotSnvTypes}
\title{Barplot of all reference contexts grouped by SNV type}
\usage{
plotSnvTypes(vcf, context)
}
\arguments{
\item{vcf}{ExpandedVcf object}

\item{context}{\code{DNAStringSet} returned by \code{snvtyping::getSnvContext} function}
}
\value{
Barplot of all reference contexts grouped by SNV type
}
\description{
This functions plots a bar plot using \code{ggplot2} of all reference contexts
grouping them by SNV type.
}
\examples{
library(snvtyping)
library(S4Vectors)
context.len <- 3
genome <- BSgenome.Hsapiens.UCSC.hg19::Hsapiens
fl <- system.file("extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
                  mustWork=TRUE)
vcf <- VariantAnnotation::readVcf(fl)
vcf <- VariantAnnotation::expand(vcf)
vcf <- preprocessVcf(vcf)
vcf <- vcf[1:50]
context <- getSnvContext(vcf, genome, context.len)
plotSnvTypes(vcf, context)
}
