% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessVcf.R
\name{preprocessVcf}
\alias{preprocessVcf}
\title{Preprocess Genome According to VCF Information}
\usage{
preprocessVcf(vcf)
}
\arguments{
\item{vcf}{\code{ExpandededVCF} object}
}
\value{
Set of relevant sequences on which variant calling was made.
Usually subset of the original genome.
}
\description{
This function will filter the genome by taking into account only relevant
chromosomes used in the variant calling experiment that generated the VCF
file in question.
}
\examples{
library(snvtyping)
library(S4Vectors)
context.len <- 3
genome <- BSgenome.Hsapiens.UCSC.hg19::Hsapiens
fl <- system.file("extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
                  mustWork=TRUE)
vcf <- VariantAnnotation::readVcf(fl)
vcf <- VariantAnnotation::expand(vcf)
vcf <- preprocessVcf(vcf)
}
