% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvtyping-package.R
\docType{package}
\name{snvtyping-package}
\alias{snvtyping}
\alias{snvtyping-package}
\title{snvtyping: Genomic Context of SNVs From VCF File}
\description{
This package can analyze VCF files containing SNV mutations and convert them in a readable Data Frame. The main feature is to get the genomic context of all Single Nucleotide Variants mutations in a VCF file; it can also classify and plot them.
}
\author{
\strong{Maintainer}: Alessandro Pandolfi \email{alessandro.pandolfi@protonmail.com}

}
\keyword{internal}
